+++ package.json
@@ -37,6 +37,7 @@
       "hand-wave",
       "octagon-exclamation",
       "thumbs-up",
+      "triangle-exclamation",
       "user-lock"
     ],
     "fas": [
