+++ src/js/outreach/apps/verify_app.js
@@ -11,10 +11,8 @@ import {
   NotAvailableView,
 } from 'js/outreach/views/verify_views';
 
-import {
-  DialogView,
-  ErrorView,
-} from 'js/outreach/views/dialog_views';
+import { DialogView } from 'js/outreach/views/dialog_views';
+import { ErrorView } from 'js/outreach/views/error_views';
 
 export default App.extend({
   beforeStart({ actionId }) {
