+++ src/js/outreach/views/error.scss
@@ -1,0 +1,14 @@
+.dialog__icon--error {
+  color: $error-color;
+  margin: 0 auto 24px;
+}
+
+.dialog__error-header {
+  font-size: 20px;
+}
+
+.dialog__error-info {
+  font-size: 16px;
+  font-weight: 600;
+  margin-top: 24px;
+}
